+++ exolve-m.js
@@ -598,7 +598,7 @@ Exolve.prototype.init = function() {
               <a id="${this.prefix}-webifi" href="" class="xlv-toggler"
                   title="${this.textLabels['webifi.hover']}"
                   >${this.textLabels['webifi']}</a>
-              <span id="${this.prefix}-copyright"></span>
+              <span class="xlv-copyright" id="${this.prefix}-copyright"></span>
               <div id="${this.prefix}-tools" class="xlv-toggleable"
                   style="display:none">
                 <p id="${this.prefix}-id" class="xlv-metadata">
@@ -813,11 +813,12 @@ Exolve.prototype.init = function() {
   } else {
     this.copyright = '';
   }
-  let smallPrintBox = document.getElementById(this.prefix + '-small-print');
+  this.smallPrint = document.getElementById(this.prefix + '-small-print');
   for (credit of this.credits) {
-    smallPrintBox.insertAdjacentHTML('beforeend',
+    this.smallPrint.insertAdjacentHTML('beforeend',
         '<div class="xlv-credit">' + credit + '</div>');
   }
+  this.controlsEtc = document.getElementById(this.prefix + '-controls-etc');
 
   this.gridPanel = document.getElementById(this.prefix + '-grid-panel');
   this.svg = document.getElementById(this.prefix + '-grid');
@@ -7442,6 +7443,13 @@ Exolve.prototype.handleBeforePrint = function() {
     document.body.insertAdjacentElement('afterbegin', this.frame);
     this.printingChanges.moves.push(
         {'elem': this.frame, 'target': puzParent, 'sibling': puzSibling});
+    if (this.copyright) {
+      const crt = document.getElementById(this.prefix + '-copyright');
+      const sibling = crt.nextSibling;
+      this.controlsEtc.insertAdjacentElement('afterbegin', crt);
+      this.printingChanges.moves.push(
+          {'elem': crt, 'target': this.smallPrint, 'sibling': sibling});
+    }
 
     const hider = document.createElement('div');
     hider.className = 'xlv-dont-print';
@@ -7614,8 +7622,7 @@ Exolve.prototype.printTwoColumns = function(settings) {
   `;
   this.frame.appendChild(customStyles);
 
-  const controlsEtcH = document.getElementById(
-      this.prefix + '-controls-etc').clientHeight;
+  const controlsEtcH = this.controlsEtc.clientHeight;
 
   // We need to apply the print media style, with an additional 2-col
   // grid layout. We'll then measure clue row heights and balance
@@ -7769,8 +7776,7 @@ Exolve.prototype.printThreeColumns = function(settings) {
   `;
   this.frame.appendChild(customStyles);
 
-  const controlsEtcH = document.getElementById(
-      this.prefix + '-controls-etc').clientHeight;
+  const controlsEtcH = this.controlsEtc.clientHeight;
 
   // We need to apply the print media style, with an additional 3-col
   // grid layout. We'll then measure clue row heights and balance